requirements.txt
# Boltz-2 SAR Iterator Requirements

# Core ML/Structural Biology
boltz[cuda]>=1.0.0  # Use boltz (CPU-only) if no GPU available

# Data handling
pandas>=1.5.0
numpy>=1.23.0

# YAML processing
pyyaml>=6.0

# Structure file handling
gemmi>=0.6.0  # For PDB to CIF conversion

# Statistical analysis and metrics
scipy>=1.9.0
scikit-learn>=1.2.0

# Optional: For visualization of results
matplotlib>=3.6.0
seaborn>=0.12.0
